[addons][base] separate now all C parts complete to addon_base.h
